backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
yt-dlp>=2024.1.0
deepgram-sdk>=6.0.0
anthropic>=0.25.0
